chore: Rename `validation.ts`
chore: Add `validation_.ts` to the barrel file

Fix the validation_ import in validation_.test.ts
