+++ src/typescript/sdk/src/utils/index.ts
@@ -10,3 +10,4 @@ export * from "./parse-changes-for-balances";
 export * from "./sum-emoji-bytes";
 export * from "./type-tags";
 export * from "./validation";
+export * from "./validation_";
