+++ src/typescript/sdk/src/utils/validation_.ts
